refactor(profiling): split lock profiling in two parts
This splits the lock profiling code to be able to profile any type of lock.
