+++ ddtrace/profiling/exporter/pprof.pyx
@@ -12,6 +12,7 @@ from ddtrace.internal.utils import config
 from ddtrace.profiling import event
 from ddtrace.profiling import exporter
 from ddtrace.profiling import recorder
+from ddtrace.profiling.collector import _lock
 from ddtrace.profiling.collector import memalloc
 from ddtrace.profiling.collector import stack_event
 from ddtrace.profiling.collector import threading
@@ -284,7 +285,7 @@ class _PprofConverter(object):
         trace_type,  # type: str
         frames,  # type: HashableStackTraceType
         nframes,  # type: int
-        events,  # type: typing.List[threading.LockAcquireEvent]
+        events,  # type: typing.List[_lock.LockAcquireEvent]
         sampling_ratio,  # type: float
     ):
         # type: (...) -> None
@@ -321,7 +322,7 @@ class _PprofConverter(object):
         trace_type,  # type: str
         frames,  # type: HashableStackTraceType
         nframes,  # type: int
-        events,  # type: typing.List[threading.LockReleaseEvent]
+        events,  # type: typing.List[_lock.LockReleaseEvent]
         sampling_ratio,  # type: float
     ):
         # type: (...) -> None
@@ -504,7 +505,7 @@ class PprofExporter(exporter.Exporter):
 
     def _lock_event_group_key(
         self,
-        event: threading.LockEventBase,
+        event: _lock.LockEventBase,
     ) -> LockEventGroupKey:
         return LockEventGroupKey(
             _none_to_str(event.lock_name),
@@ -521,8 +522,8 @@ class PprofExporter(exporter.Exporter):
         )
 
     def _group_lock_events(
-        self, events: typing.Iterable[threading.LockEventBase]
-    ) -> typing.Iterator[typing.Tuple[LockEventGroupKey, typing.Iterator[threading.LockEventBase]]]:
+        self, events: typing.Iterable[_lock.LockEventBase]
+    ) -> typing.Iterator[typing.Tuple[LockEventGroupKey, typing.Iterator[_lock.LockEventBase]]]:
         return itertools.groupby(
             sorted(events, key=self._lock_event_group_key),
             key=self._lock_event_group_key,
@@ -617,8 +618,8 @@ class PprofExporter(exporter.Exporter):
 
         # Handle Lock events
         for event_class, convert_fn in (
-            (threading.LockAcquireEvent, converter.convert_lock_acquire_event),
-            (threading.LockReleaseEvent, converter.convert_lock_release_event),
+            (_lock.LockAcquireEvent, converter.convert_lock_acquire_event),
+            (_lock.LockReleaseEvent, converter.convert_lock_release_event),
         ):
             lock_events = events.get(event_class, [])  # type: ignore[call-overload]
             sampling_sum_pct = sum(event.sampling_pct for event in lock_events)
