+++ tests/profiling/simple_program_fork.py
@@ -34,23 +34,23 @@ if child_pid == 0:
     lock.release()
 
     # We don't track it
-    assert test_lock_name not in set(e.lock_name for e in recorder.reset()[cthreading.LockReleaseEvent])
+    assert test_lock_name not in set(e.lock_name for e in recorder.reset()[cthreading.ThreadingLockReleaseEvent])
 
     # We track this one though
     lock = threading.Lock()
     test_lock_name = "simple_program_fork.py:40"
-    assert test_lock_name not in set(e.lock_name for e in recorder.reset()[cthreading.LockAcquireEvent])
+    assert test_lock_name not in set(e.lock_name for e in recorder.reset()[cthreading.ThreadingLockAcquireEvent])
     lock.acquire()
     events = recorder.reset()
-    assert test_lock_name in set(e.lock_name for e in events[cthreading.LockAcquireEvent])
-    assert test_lock_name not in set(e.lock_name for e in events[cthreading.LockReleaseEvent])
+    assert test_lock_name in set(e.lock_name for e in events[cthreading.ThreadingLockAcquireEvent])
+    assert test_lock_name not in set(e.lock_name for e in events[cthreading.ThreadingLockReleaseEvent])
     lock.release()
-    assert test_lock_name in set(e.lock_name for e in recorder.reset()[cthreading.LockReleaseEvent])
+    assert test_lock_name in set(e.lock_name for e in recorder.reset()[cthreading.ThreadingLockReleaseEvent])
 
     parent_events = parent_recorder.reset()
     # Let's sure our copy of the parent recorder does not receive it since the parent profiler has been stopped
-    assert test_lock_name not in set(e.lock_name for e in parent_events[cthreading.LockAcquireEvent])
-    assert test_lock_name not in set(e.lock_name for e in parent_events[cthreading.LockReleaseEvent])
+    assert test_lock_name not in set(e.lock_name for e in parent_events[cthreading.ThreadingLockAcquireEvent])
+    assert test_lock_name not in set(e.lock_name for e in parent_events[cthreading.ThreadingLockReleaseEvent])
 
     # This can run forever if anything is broken!
     while not recorder.events[stack_event.StackSampleEvent]:
@@ -58,9 +58,9 @@ if child_pid == 0:
 else:
     recorder = ddtrace.profiling.bootstrap.profiler._profiler._recorder
     assert recorder is parent_recorder
-    assert test_lock_name not in set(e.lock_name for e in recorder.reset()[cthreading.LockReleaseEvent])
+    assert test_lock_name not in set(e.lock_name for e in recorder.reset()[cthreading.ThreadingLockReleaseEvent])
     lock.release()
-    assert test_lock_name in set(e.lock_name for e in recorder.reset()[cthreading.LockReleaseEvent])
+    assert test_lock_name in set(e.lock_name for e in recorder.reset()[cthreading.ThreadingLockReleaseEvent])
     assert ddtrace.profiling.bootstrap.profiler.status == service.ServiceStatus.RUNNING
     print(child_pid)
     pid, status = os.waitpid(child_pid, 0)
